I think this should work, but it's not
