+++ backend/resolvers/badgeResolver.ts
@@ -5,20 +5,61 @@ import {
   Badge,
   BadgeLevel,
 } from "@prisma/client";
-import BadgeService from "../services/implementation/badgeImplementation";
-import IBadgeService from "../services/interface/badgeInterface";
 
-const badgeService: IBadgeService = new BadgeService();
+import { getNow } from "../utils/formatDateTime";
+
+const prisma = new PrismaClient();
 
 const badgeResolver = {
   Query: {
     getCustomBadges: async (): Promise<Badge[]> => {
-      return badgeService.getCustomBadges();
+      try {
+        const customBadges = await prisma.badge.findMany({
+          where: {
+            badge_type: "CUSTOM",
+          },
+          include: {
+            badge_level: true,
+          },
+          orderBy: {
+            name: 'asc'
+          },
+        });
+        return customBadges;
+      } catch (err) {
+        if (err instanceof Error) {
+          throw new Error(err.message || "Failed to get custom badges.");
+        }
+        throw new Error("Failed to get custom badges.");
+      }
     },
     getSystemBadges: async (): Promise<Badge[]> => {
-      return badgeService.getSystemBadges();
+        try {
+        const systemBadges = await prisma.badge.findMany({
+          where: {
+            badge_type: "SYSTEM",
+          },
+          include: {
+            badge_level: {
+              orderBy: {
+                level: 'asc',
+              },
+            },
+          },
+          orderBy: {
+            name: 'asc'
+          },
+        });
+        return systemBadges;
+      } catch (err) {
+        if (err instanceof Error) {
+          throw new Error(err.message || "Failed to get custom badges.");
+        }
+        throw new Error("Failed to get custom badges.");
+      }
     },
   },
+
   Mutation: {
     updateBadgeStatus: async (
       _parent: undefined,
@@ -30,8 +71,26 @@ const badgeResolver = {
         is_active: boolean,
       }
     ): Promise<boolean> => {
-      return badgeService.updateBadgeStatus(badge_id, is_active)
+      const badge = await prisma.badge.findUnique({
+        where: { badge_id },
+      });
+      if (!badge){
+        throw new Error("Badge not found");
+      }
+      try {
+        await prisma.badge.update({
+          where: { badge_id },
+          data: {
+            is_active
+          },
+        });
+        return true;
+      } catch (err) {
+        // @ts-ignore
+        throw new Error(err.message || "Failed to update badge");
+      }
     },
+
     assignCustomBadge: async (
       _parent: undefined,
       {
@@ -44,12 +103,72 @@ const badgeResolver = {
         participant_ids: number[];
       }
     ): Promise<number[]> => {
-      return badgeService.assignCustomBadge(
-        badge_id,
-        marillac_bucks,
-        participant_ids
+      const badge = await prisma.badge.findUnique({
+        where: { badge_id },
+      });
+      if (!badge) throw new Error(`Badge with ID ${badge_id} not found`);
+
+      const participants = await prisma.participant.findMany({
+        where: {
+          participant_id: {
+            in: participant_ids,
+          },
+        },
+      });
+      if (participants.length !== participant_ids.length)
+        throw new Error(`Some IDs were invalid!`);
+
+      const existingBadges = await prisma.earnedBadge.findMany({
+        where: {
+          participant_id: { in: participant_ids },
+          name: badge.name,
+        },
+        select: { participant_id: true },
+      });
+      const alreadyEarnedIds = new Set(
+        existingBadges.map((b) => b.participant_id)
       );
+      const eligibleParticipantIds = participant_ids.filter(
+        (id) => !alreadyEarnedIds.has(id)
+      );
+      
+      if (eligibleParticipantIds.length === 0) {
+        throw new Error("Participants have already received this badge");
+      }
+
+      await prisma.$transaction(
+        eligibleParticipantIds.map((id) =>
+          prisma.earnedBadge.create({
+            data: {
+              badge_id: badge.badge_id,
+              participant_id: id,
+              date_received: new Date().toISOString(),
+              name: badge.name,
+              description: badge.description,
+              badge_icon: badge.icon,
+              level: 1,
+            },
+          }),
+        )
+      );
+      
+      // also add marillac bucks for the earned badge
+      await prisma.$transaction(
+        eligibleParticipantIds.map((id) =>
+          prisma.participant.update({
+            where: { participant_id: id },
+            data: {
+              marillac_bucks: {
+                increment: marillac_bucks,
+              },
+            },
+          }),
+        )
+      );
+      
+      return eligibleParticipantIds;
     },
+
     editCustomBadge: async (
       _parent: undefined,
       {
@@ -62,11 +181,25 @@ const badgeResolver = {
         new_custom_badge_description?: string;
       }
     ): Promise<boolean> => {
-      return badgeService.editCustomBadge(
-        custom_badge_id,
-        new_custom_badge_name,
-        new_custom_badge_description
-      );
+        try {
+        if (new_custom_badge_name == undefined && new_custom_badge_description == undefined)
+          throw new Error("No edits provided");
+        await prisma.badge.update({
+          where: { badge_id: custom_badge_id, badge_type: "CUSTOM" },
+          data: {
+            ...(new_custom_badge_name !== undefined && {
+              name: new_custom_badge_name,
+            }),
+            ...(new_custom_badge_description !== undefined && {
+              description: new_custom_badge_description,
+            }),
+          },
+        });
+        return true;
+      } catch (err) {
+        // @ts-ignore
+        throw new Error(err.message || "Something went wrong");
+      }
     },
     editSystemBadge: async (
       _parent: undefined,
@@ -80,12 +213,27 @@ const badgeResolver = {
         system_badge_criteria?: string;
       }
     ): Promise<boolean> => {
-      return badgeService.editSystemBadge(
-        system_badge_id,
-        system_badge_name,
-        system_badge_criteria
-      );
+      const badge = await prisma.badge.findUnique({
+        where: {badge_id:system_badge_id},
+      });
+      if (!badge || badge.badge_type !== "SYSTEM"){
+        throw new Error("BAdge not found or not a system badge");
+      }
+      try {
+        await prisma.badge.update({
+          where: { badge_id: system_badge_id},
+          data: {
+            name: system_badge_name,
+            ...(system_badge_criteria !== undefined && { description: system_badge_criteria }),
+          },
+        });
+        return true;
+      } catch (err) {
+        // @ts-ignore
+        throw new Error(err.message || "Failed to update system badge");
+      }
     },
+    
     createCustomBadge: async (
       _parent: undefined,
       {
@@ -98,11 +246,20 @@ const badgeResolver = {
         icon: Icon;
       }
     ): Promise<boolean> => {
-      return badgeService.createCustomBadge(
-        name,
-        description,
-        icon
-      );
+      try {
+        await prisma.badge.create({
+          data: {
+            name,
+            description,
+            icon,
+            is_consecutive: false,
+            badge_type: "CUSTOM",
+          },
+        });
+        return true;
+      } catch (err) {
+        throw new Error("Something went wrong");
+      }
     },
     deleteCustomBadge: async (
       _parent: undefined,
@@ -112,7 +269,30 @@ const badgeResolver = {
         badge_id: number;
       }
     ): Promise<boolean> => {
-      return badgeService.deleteCustomBadge(badge_id);
+        try {
+        const badge = await prisma.badge.findUnique({
+          where: { badge_id },
+        });
+
+        if (!badge) {
+          throw new Error(`Badge with ID ${badge_id} does not exist`);
+        }
+
+        if (badge.badge_type !== BadgeType.CUSTOM) {
+          throw new Error(
+            `Badge with ID ${badge_id} is not a custom badge and cannot be deleted`
+          );
+        }
+
+        await prisma.badge.delete({
+          where: { badge_id },
+        });
+
+        return true;
+      } catch (err) {
+        console.error(`Failed to delete badge ${badge_id}:`, err);
+        return false;
+      }
     },
     editBadgeLevel: async (
       _parent: undefined,
@@ -128,12 +308,24 @@ const badgeResolver = {
         marillac_bucks: number;
       }
     ): Promise<boolean> => {
-      return badgeService.editBadgeLevel(
-        badge_id,
-        badge_level,
-        benchmark,
-        marillac_bucks
-      );
+      try {
+        await prisma.badgeLevel.update({
+          where: {
+            badge_id_level: {
+              badge_id,
+              level: badge_level,
+            },
+          },
+          data: {
+            benchmark,
+            marillac_bucks,
+          },
+        });
+        return true;
+      } catch (err) {
+        // @ts-ignore
+        throw new Error(err.message || "Something went wrong");
+      }
     },
   },
 };
